update install_defender_gcp_marketplace.adoc
The first image in this document is outdated - the search parameter used to get to the Prisma Cloud Kubernetes Security Defender that shows the referenced "Configure" button is by searching "Prisma Cloud - Kubernetes Security Defender" . The image shows an outdated version of the application, that does not have the configure button, and will take customers to the wrong part of the setup.

I have a picture of the correct version available, but I am unsure how to add it here. I can present it upon request.