update install_defender_gcp_marketplace.adoc
The image for the first image of the Prisma Cloud - Kubernetes Security Defender is out-dated. The correct image can be snipped from searching on the GCP marketplace for "Prisma Cloud - Kubernetes Security Defender" on the GCP marketplace.

Also, the search term in the first paragraph should be edited to reflect the above.
I can supply the .png at request- but can't figure out how to get it in here at the moment.